// frontend/web/angular-admin/src/app/generated/model/boutique.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Boutique { 
    boutiqueDescription?: string;
    boutiqueImage?: Blob;
    boutiqueName?: string;
    boutiqueQuater?: string;
    boutiqueUser?: string;
}