// frontend/web/angular-admin/src/app/generated/model/uploadResponse.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UploadResponse { 
    resourceUrl?: string;
    uploadDate?: Date;
    uploadLocation?: string;
}